backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
google-cloud-aiplatform==1.75.0
google-cloud-speech==2.21.0
